map searcher: add location searcher component
