+++ frontend/containers/discover-map/location-searcher/index.ts
@@ -1,0 +1,2 @@
+export type { LocationSearcherProps } from './types';
+export { default } from './component';
