+++ frontend/containers/discover-map/location-searcher/types.ts
@@ -1,0 +1,1 @@
+export interface LocationSearcherProps {}
